# api/openapi.yaml
openapi: 3.0.1
info:
  description: "TODO: Add a description"
  title: FIWARE-NGSI v2 Specification
  version: "1.0"
servers:
- url: http://orion.lab.fiware.org/
paths:
...
  /v2/entities/{entityId}/attrs/{attrName}:
    delete:
      description: |-
        Removes an entity attribute.
        Response:
        * Successful operation uses 204 No Content
        * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
          more details.
      operationId: Remove a Single Attribute
      parameters:
      - description: Id of the entity.
        in: path
        name: entityId
        required: true
        schema:
          type: string
      - description: Attribute name.
        in: path
        name: attrName
        required: true
        schema:
          type: string
      - description: |-
          Entity type, to avoid ambiguity in case there are several
          entities with the same entity id.
        in: query
        name: type
        schema:
          type: string
      responses:
        "204":
          content: {}
          description: ""
      summary: Remove a Single Attribute
      tags:
      - Attributes
      x-unitTests:
      - request:
          method: DELETE
          uri: /v2/entities/Bcn_Welt/attrs/temperature
        expectedResponse:
          x-allowExtraHeaders: true
          x-bodyMatchMode: NONE
          x-arrayOrderedMatching: false
          x-arrayCheckCount: false
          x-matchResponseSchema: true
          headers: {}
        x-testShouldPass: true
        x-testEnabled: true
        x-testName: Remove a Single Attribute1
        x-testDescription: |-
          Removes an entity attribute.
          Response:
          * Successful operation uses 204 No Content
          * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
            more details.
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-accepts:
      - application/json
    get:
      description: |-
        Returns a JSON object with the attribute data of the attribute. The object follows the JSON
        representation for attributes (described in "JSON Attribute Representation" section).
        Response:
        * Successful operation uses 200 OK.
        * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
          more details.
      operationId: Get attribute data
      parameters:
      - description: Id of the entity
        in: path
        name: entityId
        required: true
        schema:
          type: string
      - description: Name of the attribute to be retrieved.
        in: path
        name: attrName
        required: true
        schema:
          type: string
      - description: |-
          Entity type, to avoid ambiguity in case there are several
          entities with the same entity id.
        in: query
        name: type
        schema:
          type: string
      - description: |-
          A list of metadata names to include in the response.
          See "Filtering out attributes and metadata" section for more detail.
        in: query
        name: metadata
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              example:
                value: 21.7
                type: Number
                metadata: {}
              schema:
                $ref: '#/components/schemas/GetAttributeDataResponse'
          description: ""
      summary: Get attribute data
      tags:
      - Attributes
      x-unitTests:
      - request:
          method: GET
          uri: /v2/entities/Bcn_Welt/attrs/temperature?metadata=accuracy
        expectedResponse:
          x-allowExtraHeaders: true
          x-bodyMatchMode: RAW
          x-arrayOrderedMatching: false
          x-arrayCheckCount: false
          x-matchResponseSchema: true
          headers:
            Content-Type: application/json
          body: "{  \"value\": 21.7,  \"type\": \"Number\",  \"metadata\": {}}"
        x-testShouldPass: true
        x-testEnabled: true
        x-testName: Get attribute data1
        x-testDescription: |-
          Returns a JSON object with the attribute data of the attribute. The object follows the JSON
          representation for attributes (described in "JSON Attribute Representation" section).
          Response:
          * Successful operation uses 200 OK.
          * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
            more details.
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-accepts:
      - application/json
    put:
      description: |-
        The request payload is an object representing the new attribute data. Previous attribute data
        is replaced by the one in the request. The object follows the JSON representation for attributes
        (described in "JSON Attribute Representation" section).
        Response:
        * Successful operation uses 204 No Content
        * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
          more details.
      operationId: Update Attribute Data
      parameters:
      - description: Id of the entity to update
        in: path
        name: entityId
        required: true
        schema:
          type: string
      - description: Attribute name
        in: path
        name: attrName
        required: true
        schema:
          type: string
      - in: header
        name: Content-Type
        required: true
        schema:
          type: string
      - description: |-
          Entity type, to avoid ambiguity in case there are several
          entities with the same entity id.
        in: query
        name: type
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAttributeDataRequest'
        required: true
      responses:
        "200":
          content: {}
          description: ""
      summary: Update Attribute Data
      tags:
      - Attributes
      x-unitTests:
      - request:
          method: PUT
          uri: /v2/entities/Bcn_Welt/attrs/temperature
          headers:
            Content-Type: application/json
          body: "{  \"value\": 25.0,  \"metadata\": {    \"unitCode\": {      \"value\"\
            : \"CEL\"    }  }}"
        expectedResponse:
          x-allowExtraHeaders: true
          x-bodyMatchMode: NONE
          x-arrayOrderedMatching: false
          x-arrayCheckCount: false
          x-matchResponseSchema: true
          headers: {}
        x-testShouldPass: true
        x-testEnabled: true
        x-testName: Update Attribute Data1
        x-testDescription: |-
          The request payload is an object representing the new attribute data. Previous attribute data
          is replaced by the one in the request. The object follows the JSON representation for attributes
          (described in "JSON Attribute Representation" section).
          Response:
          * Successful operation uses 204 No Content
          * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
            more details.
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
...
    GetAttributeDataResponse:
      example:
        value: 21.7
        type: Number
        metadata: {}
      properties:
        value:
          description: ""
          example: 21.7
          format: double
          type: number
        type:
          description: ""
          example: Number
          type: string
        metadata:
          description: ""
          example: {}
          properties: {}
          type: object
      required:
      - metadata
      - type
      - value
      title: Get attribute data response
      type: object
    UpdateAttributeDataRequest:
      example:
        value: 25.0
        metadata:
          unitCode:
            value: CEL
      properties:
        value:
          description: ""
          example: 25.0
          format: double
          type: number
        metadata:
          description: ""
          example:
            unitCode:
              value: CEL
          properties: {}
          type: object
      required:
      - metadata
      - value
      title: Update Attribute Data request
      type: object
...